toolkit/devtools/webconsole/test/test_consoleapi.html

<!DOCTYPE HTML>
<html lang="en">
<head>
  <meta charset="utf8">
  <title>Test for the Console API</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript;version=1.8" src="common.js"></script>
  <!-- Any copyright is dedicated to the Public Domain.
     - http://creativecommons.org/publicdomain/zero/1.0/ -->
</head>
<body>
<p>Test for the Console API</p>

<script type="text/javascript;version=1.8">
netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
SimpleTest.waitForExplicitFinish();

let expectedConsoleCalls = [];

function doConsoleCalls(aState)
{
  console.log("foobarBaz-log", undefined);
  console.info("foobarBaz-info", null);
  console.warn("foobarBaz-warn", document.body);
  console.debug(null);
  console.trace();
  console.dir(document, window);

  expectedConsoleCalls = [
    {
      level: "log",
      filename: /test_consoleapi/,
      functionName: "doConsoleCalls",
      timeStamp: /^\d+$/,
      arguments: ["foobarBaz-log", { type: "undefined" }],
    },
    {
      level: "info",
      filename: /test_consoleapi/,
      functionName: "doConsoleCalls",
      timeStamp: /^\d+$/,
      arguments: ["foobarBaz-info", { type: "null" }],
    },
    {
      level: "warn",
      filename: /test_consoleapi/,
      functionName: "doConsoleCalls",
      timeStamp: /^\d+$/,
      arguments: ["foobarBaz-warn", { type: "object", actor: /[a-z]/ }],
    },
    {
      level: "debug",
      filename: /test_consoleapi/,
      functionName: "doConsoleCalls",
      timeStamp: /^\d+$/,
      arguments: [{ type: "null" }],
    },
    {
      level: "trace",
      filename: /test_consoleapi/,
      functionName: "doConsoleCalls",
      timeStamp: /^\d+$/,
      arguments: [
        {
          filename: /test_consoleapi/,
          functionName: "doConsoleCalls",
        },
        {
          filename: /test_consoleapi/,
          functionName: "onAttach",
        },
      ],
    },
    {
      level: "dir",
      filename: /test_consoleapi/,
      functionName: "doConsoleCalls",
      timeStamp: /^\d+$/,
      arguments: [
        {
          type: "object",
          actor: /[a-z]/,
          className: "HTMLDocument",
        },
        {
          type: "object",
          actor: /[a-z]/,
          className: "Window",
        }
      ],
      objectProperties: [
        {
          name: "ATTRIBUTE_NODE",
          value: 2,
        },
        {
          name: "CDATA_SECTION_NODE",
          value: 4,
        }, // ...
      ],
    },
  ];
}

function startTest()
{
  removeEventListener("load", startTest);

  attachConsole(["ConsoleAPI"], onAttach);
}

function onAttach(aState, aResponse)
{
  onConsoleAPICall = onConsoleAPICall.bind(null, aState);
  aState.dbgClient.addListener("consoleAPICall", onConsoleAPICall);
  doConsoleCalls(aState.actor);
}

let consoleCalls = [];

function onConsoleAPICall(aState, aType, aPacket)
{
  is(aPacket.from, aState.actor, "console API call actor");

  consoleCalls.push(aPacket.message);
  if (consoleCalls.length != expectedConsoleCalls.length) {
    return;
  }

  aState.dbgClient.removeListener("consoleAPICall", onConsoleAPICall);

  expectedConsoleCalls.forEach(function(aMessage, aIndex) {
    info("checking received console call #" + aIndex);
    checkConsoleAPICall(consoleCalls[aIndex], expectedConsoleCalls[aIndex]);
  });


  consoleCalls = [];

  closeDebugger(aState, function() {
    SimpleTest.finish();
  });
}

addEventListener("load", startTest);
</script>
</body>
</html>
